[IMP] mail: move getters partition() and actions() in standard actions
Feature was specific to thread and composer actions. This is moved
in standard action so that all actions support groups.
